+++ scripts/embedding/gprolog/gprolog_embed_logtalk.sh
@@ -2,10 +2,10 @@
 
 #############################################################################
 ## 
-##   This script creates a new GNU Prolog top-level
-##   interpreter that embeds Logtalk
+##   This script creates a new GNU Prolog top-level interpreter
+##   that embeds Logtalk and optionally a Logtalk application
 ## 
-##   Last updated on April 7, 2018
+##   Last updated on April 12, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -27,6 +27,7 @@
 
 directory="$HOME/collect"
 paths="$LOGTALKHOME/paths/paths_core.pl"
+compile="false"
 
 if ! [ "$LOGTALKHOME" ]; then
 	echo "The environment variable LOGTALKHOME should be defined first, pointing"
@@ -66,7 +67,7 @@ elif ! [ -d "$LOGTALKHOME" ]; then
 fi
 
 print_version() {
-	echo "$(basename "$0") 0.6"
+	echo "$(basename "$0") 0.7"
 	exit 0
 }
 
@@ -78,29 +79,33 @@ usage_help()
 	echo "source code given its loader file."
 	echo
 	echo "Usage:"
-	echo "  $(basename "$0") [-d directory] [-p paths] [-l loader] [-s settings]"
+	echo "  $(basename "$0") [-c] [-d directory] [-p paths] [-h hooks] [-s settings] [-l loader]"
 	echo "  $(basename "$0") -v"
 	echo "  $(basename "$0") -h"
 	echo
 	echo "Optional arguments:"
-	echo "  -v print version of $(basename "$0")"
+	echo "  -c compile library alias paths in paths and settings files"
 	echo "  -d directory to use for intermediate and final results (default is $directory)"
 	echo "  -p library paths file (default is $paths)"
+	echo "  -h YAP hooks file (default is $hooks)"
+	echo "  -s optional settings file"
 	echo "  -l optional loader file for the application"
-	echo "  -s optional settings file for the application"
+	echo "  -v print version of $(basename "$0")"
 	echo "  -h help"
 	echo
 	exit 0
 }
 
-while getopts "vd:p:l:s:h" option
+while getopts "cd:p:h:l:s:vh" option
 do
 	case $option in
-		v) print_version;;
+		c) compile="true";;
 		d) d_arg="$OPTARG";;
 		p) p_arg="$OPTARG";;
+		h) h_arg="$OPTARG";;
 		l) l_arg="$OPTARG";;
 		s) s_arg="$OPTARG";;
+		v) print_version;;
 		h) usage_help;;
 		*) usage_help;;
 	esac
@@ -156,14 +161,23 @@ else
 	extension=''
 fi
 
-gplgt$extension --query-goal "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$directory')]),halt"
-
 cp "$LOGTALKHOME/adapters/gnu.pl" .
-cp "$paths" ./paths.pl
 cp "$LOGTALKHOME/core/core.pl" .
 
+gplgt$extension --query-goal "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$directory')]),halt"
+
+if [ "$compile" != "false" ] ; then
+	gplgt$extension --query-goal "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$paths',[hook(expand_library_alias_paths),scratch_directory('$directory')]),halt"
+else
+	cp "$paths" "$directory/paths_lgt.pl"
+fi
+
 if [ "$settings" != "" ] ; then
-	gplgt$extension --query-goal "logtalk_compile('$settings',[optimize(on),scratch_directory('$directory')]),halt"
+	if [ "$compile" != "false" ] ; then
+		gplgt$extension --query-goal "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$settings',[hook(expand_library_alias_paths),optimize(on),scratch_directory('$directory')]),halt"
+	else
+		gplgt$extension --query-goal "logtalk_compile('$settings',[optimize(on),scratch_directory('$directory')]),halt"
+	fi
 else
 	touch settings_lgt.pl
 fi
@@ -180,7 +194,7 @@ else
 	touch application.pl
 fi
 
-gplc -o logtalk gnu.pl expanding*_lgt.pl monitoring*_lgt.pl forwarding*_lgt.pl user*_lgt.pl logtalk*_lgt.pl core_messages*_lgt.pl $(ls application/*.pl) core.pl settings*_lgt.pl paths.pl
+gplc -o logtalk gnu.pl expanding*_lgt.pl monitoring*_lgt.pl forwarding*_lgt.pl user*_lgt.pl logtalk*_lgt.pl core_messages*_lgt.pl $(ls application/*.pl) core.pl settings*_lgt.pl paths_*.pl
 
 rm *.pl
 rm application/*.pl